requirements.txt
Flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
openai==1.6.1
python-dotenv==1.0.0

